refactor(dispatch): extract pure error classification to dispatch-failures.ts
First step of the dispatch.ts split, done by RISK: dispatch.ts is the audit's most dangerous
file (6 module-level singletons whose misplacement silently breaks the .dispatch-state.json
mutex, an import cycle needing DI surgery, a cross-package ledger contract on the state file),
so the pure, zero-state modules come out first and the fire-path/singleton/cycle modules are
left for a dedicated, rig-E2E'd effort.

Moves the error-classification layer — formatDispatchError, classifyDispatchError,
classifyDispatchFailureDetail, formatSanitizedRuntimeFailure, and their types — into
src/core/dispatch-failures.ts. These are pure functions over the adapter's typed
RuntimeError/RuntimeTurnError with zero dispatch state. dispatch.ts re-exports the public
classify* surface so every `@/core/dispatch` consumer and the dispatch-error-classification
test keep their import path unchanged. dispatch.ts: 2,240 → 2,096.

Verified: typecheck + lint clean; dispatch tests 61/0 (incl. the error-classification suite
through the re-export); full suite 5105/0; full 3-target binary build.
